requirements.txt
google-generativeai
streamlit
python-dotenv
pdfplumber
bs4